+++ VKapp/VKapp/NetworkService/VKAPIService.swift
@@ -1,0 +1,54 @@
+// VKAPIService.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Alamofire
+import Foundation
+
+/// Network requests
+final class VKAPIService {
+    // MARK: - Constants
+
+    private enum Constants {
+        static let scheme = "https"
+        static let host = "oauth.vk.com"
+        static let path = "/authorize"
+        static let urlQueryItemClientName = "client_id"
+        static let urlQueryItemClientValue = "51484462"
+        static let urlQueryItemDisplayName = "display"
+        static let urlQueryItemDisplayValue = "mobile"
+        static let urlQueryItemRedirectUriName = "redirect_uri"
+        static let urlQueryItemRedirectUriValue = "https://oauth.vk.com/blank.html"
+        static let urlQueryItemScopeName = "scope"
+        static let urlQueryItemScopeValue = "262150"
+        static let urlQueryItemResponseTypeName = "response_type"
+        static let urlQueryItemResponseTypeValue = "token"
+        static let urlQueryItemVName = "v"
+        static let urlQueryItemVValue = "5.131"
+    }
+
+    // MARK: - Public Methods
+
+    func fetchData(urlString: String) {
+        AF
+            .request(urlString).responseJSON { response in
+                guard let value = response.value else { return }
+                print(value)
+            }
+    }
+
+    func createURLToLoadWebView() -> URL? {
+        var urlComponents = URLComponents()
+        urlComponents.scheme = Constants.scheme
+        urlComponents.host = Constants.host
+        urlComponents.path = Constants.path
+        urlComponents.queryItems = [
+            URLQueryItem(name: Constants.urlQueryItemClientName, value: Constants.urlQueryItemClientValue),
+            URLQueryItem(name: Constants.urlQueryItemDisplayName, value: Constants.urlQueryItemDisplayValue),
+            URLQueryItem(name: Constants.urlQueryItemRedirectUriName, value: Constants.urlQueryItemRedirectUriValue),
+            URLQueryItem(name: Constants.urlQueryItemScopeName, value: Constants.urlQueryItemScopeValue),
+            URLQueryItem(name: Constants.urlQueryItemResponseTypeName, value: Constants.urlQueryItemResponseTypeValue),
+            URLQueryItem(name: Constants.urlQueryItemVName, value: Constants.urlQueryItemVValue)
+        ]
+        return urlComponents.url
+    }
+}
